+++ src/views/UserPage.js
@@ -2,24 +2,7 @@ import { useCallback, useEffect, useState } from "react";
 import SpotifyConnect from "../components/sections/SpotifyConnect";
 import SpotifyWebApiPlayer from "../services/api/SpotifyWebApiPlayer";
 import useWebPlaybackState from "../hooks/useWebPlaybackState";
-
-function huminizeMS(ms) {
-  var seconds = (ms / 1000).toFixed(0);
-  var minutes = Math.floor(seconds / 60);
-  var hours = "";
-  if (minutes > 59) {
-    hours = Math.floor(minutes / 60);
-    hours = hours >= 10 ? hours : "0" + hours;
-    minutes = minutes - hours * 60;
-    minutes = minutes >= 10 ? minutes : "0" + minutes;
-  }
-  seconds = Math.floor(seconds % 60);
-  seconds = seconds >= 10 ? seconds : "0" + seconds;
-  if (hours !== "") {
-    return hours + ":" + minutes + ":" + seconds;
-  }
-  return minutes + ":" + seconds;
-}
+import ProgressBar from "../components/ProgressBar";
 
 function UserPage() {
   const queryParams = new URLSearchParams(window.location.search);
@@ -74,16 +57,11 @@ function UserPage() {
         handleWebPlaybackState={handleWebPlaybackState}
       />
       {webPlaybackState && (
-        <div className="track">
-          <span className="track__time">{huminizeMS(parseInt(position))}</span>
-          <span className="track__bar-time">
-            <span
-              className="track__bar-time__progress"
-              style={{ width: `${progress}%` }}
-            ></span>
-          </span>
-          <span className="track__time">{huminizeMS(parseInt(duration))}</span>
-        </div>
+        <ProgressBar
+          position={position}
+          duration={duration}
+          progress={progress}
+        />
       )}
     </>
   );
